# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Renewal Radar API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List workspaces for current user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a new workspace
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceBody"
      responses:
        "201":
          description: Created workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get a workspace by ID
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/WorkspaceIdParam"
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
    put:
      operationId: updateWorkspace
      tags: [workspaces]
      summary: Update a workspace
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/WorkspaceIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceBody"
      responses:
        "200":
          description: Updated workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"

  # ─── Workspace Members ────────────────────────────────────────────────────────
  /workspaces/{workspaceId}/members:
    get:
      operationId: listWorkspaceMembers
      tags: [users]
      summary: List members of a workspace
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/WorkspaceIdParam"
      responses:
        "200":
          description: List of workspace members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceMember"
    post:
      operationId: inviteWorkspaceMember
      tags: [users]
      summary: Invite a member to the workspace
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/WorkspaceIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberBody"
      responses:
        "201":
          description: Invited member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceMember"
  /workspaces/{workspaceId}/members/{memberId}:
    delete:
      operationId: removeWorkspaceMember
      tags: [users]
      summary: Remove a member from workspace
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/WorkspaceIdParam"
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Member removed

  # ─── Obligations ─────────────────────────────────────────────────────────────
...
  /obligations/{obligationId}/reminder-rules:
    get:
      operationId: listReminderRules
      tags: [reminders]
      summary: List reminder rules for an obligation
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ObligationIdParam"
      responses:
        "200":
          description: List of reminder rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReminderRule"
    post:
      operationId: createReminderRule
      tags: [reminders]
      summary: Create a reminder rule
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ObligationIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReminderRuleBody"
      responses:
        "201":
          description: Created reminder rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderRule"
  /obligations/{obligationId}/reminder-rules/{ruleId}:
    put:
      operationId: updateReminderRule
      tags: [reminders]
      summary: Update a reminder rule
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ObligationIdParam"
        - $ref: "#/components/parameters/RuleIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReminderRuleBody"
      responses:
        "200":
          description: Updated reminder rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderRule"
    delete:
      operationId: deleteReminderRule
      tags: [reminders]
      summary: Delete a reminder rule
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ObligationIdParam"
        - $ref: "#/components/parameters/RuleIdParam"
      responses:
        "204":
          description: Deleted

  # ─── Delivery History ────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    WorkspaceIdParam:
      name: workspaceId
      in: path
      required: true
      schema:
        type: integer
    ObligationIdParam:
      name: obligationId
      in: path
      required: true
      schema:
        type: integer
    RuleIdParam:
      name: ruleId
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Workspace:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, slug, createdAt, updatedAt]
    CreateWorkspaceBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
      required: [name]
    WorkspaceMember:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        clerkUserId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [owner, admin, member]
        joinedAt:
          type: string
          format: date-time
      required: [id, workspaceId, clerkUserId, email, role, joinedAt]
    InviteMemberBody:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, member]
          default: member
      required: [email]

    # ── Obligations ────────────────────────────────────────────────────────────
...
    ReminderRule:
      type: object
      properties:
        id:
          type: integer
        obligationId:
          type: integer
        daysBefore:
          type: integer
        channel:
          type: string
          enum: [email, in_app]
        recipientType:
          type: string
          enum: [owner, backup_owner, all_members, custom_email]
        customEmail:
          type: string
          nullable: true
        isActive:
          type: boolean
        lastTriggeredAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - obligationId
        - daysBefore
        - channel
        - recipientType
        - isActive
        - createdAt
    CreateReminderRuleBody:
      type: object
      properties:
        daysBefore:
          type: integer
          minimum: 0
        channel:
          type: string
          enum: [email, in_app]
        recipientType:
          type: string
          enum: [owner, backup_owner, all_members, custom_email]
        customEmail:
          type: string
          nullable: true
        isActive:
          type: boolean
          default: true
      required: [daysBefore, channel, recipientType]

    # ── Delivery History ───────────────────────────────────────────────────────
...